refactor logger, imgui setup, force frame update for apply button
